WIP: Refactored test_hit_counter_player into test_player.py (Need to double check incase it belongs in test_computer.py)
